# Improve condition processing