--- a/src/cfnlint/context/conditions/_condition.py
+++ b/src/cfnlint/context/conditions/_condition.py
@@ -20,19 +20,21 @@ from cfnlint.helpers import FUNCTION_CONDITIONS, is_function
 class Condition:
     instance: Any = field(init=True)
     status: bool | None = field(init=True, default=None)
-    hash: str = field(init=False)
+    hash: str = field(init=True, default="")  # Changed to init=True with default
 
     fn_equals: Equal | None = field(init=True, default=None)
     condition: list["Condition"] | "Condition" | None = field(init=True, default=None)
     cnf: BooleanFunction = field(init=True, default_factory=BooleanFunction)
 
-    def __post_init__(self):
-        object.__setattr__(self, "hash", get_hash(self.instance))
+    # Removed __post_init__ since hash is now passed in
 
     @classmethod
     def create_from_instance(
         cls, instance: Any, all_conditions: dict[str, Any]
     ) -> "Condition":
+        # Calculate hash here once
+        instance_hash = get_hash(instance)
+
         fn_k, fn_v = is_function(instance)
         if fn_k is None:
             raise ValueError("Condition value must be an object of length 1")
@@ -41,7 +43,12 @@ class Condition:
                 raise ValueError(f"{fn_v!r} value should be an array")
             if fn_k == "Fn::Equals":
                 equal = Equal.create_from_instance(fn_v)
-                return cls(instance=instance, fn_equals=equal, cnf=equal.cnf)
+                return cls(
+                    instance=instance,
+                    fn_equals=equal,
+                    cnf=equal.cnf,
+                    hash=instance_hash,
+                )
 
             condition = []
             for v in fn_v:
@@ -59,7 +66,9 @@ class Condition:
                     )
                 cnf = Not(condition[0].cnf)
 
-            return cls(instance=instance, condition=condition, cnf=cnf)
+            return cls(
+                instance=instance, condition=condition, cnf=cnf, hash=instance_hash
+            )
 
         if fn_k == "Condition":
             if not isinstance(fn_v, str):
@@ -73,7 +82,7 @@ class Condition:
                 c = Condition.create_from_instance(
                     {"Fn::Equals": [None, None]}, all_conditions
                 )
-            return cls(instance=instance, condition=c, cnf=c.cnf)
+            return cls(instance=instance, condition=c, cnf=c.cnf, hash=instance_hash)
 
         raise ValueError(f"Unknown key {fn_k!r} in condition")
 
@@ -88,6 +97,7 @@ class Condition:
             instance=self.instance,
             status=status,
             cnf=self.cnf,
+            hash=self.hash,  # Pass the hash along
         )
 
     @property