*Issue #, if available:*

*Description of changes:*
## Summary

This PR introduces significant performance optimizations to the condition processing in cfn-lint, focusing on reducing redundant hash calculations. By implementing LRU caching for satisfiability results and optimizing how condition hashes are handled, we've achieved a substantial performance improvement for complex templates.

## Problem

Profiling identified several performance bottlenecks in condition processing:

1. Redundant satisfiability checks for the same CNF expressions
2. Excessive hash calculations during condition evolution
3. Inefficient handling of condition status changes

## Changes

This PR includes two main optimizations:

### 1. OrderedDict-based LRU Cache for Satisfiability Results

• Added an OrderedDict-based cache with LRU behavior for satisfiability results
• Implemented size-limited cache (10,000 entries) to prevent memory issues
• Added helper functions _get_from_satisfiable_cache and _add_to_satisfiable_cache

### 2. Optimized Condition Hash Handling

• Changed the hash field in the Condition class from init=False to init=True with a default value
• Removed the __post_init__ method that was calculating the hash for every condition
• Calculated the hash once in create_from_instance and passed it along
• Updated the evolve method to pass the hash along

## Performance Improvements

| Metric | Before (main branch) | After | Improvement |
|--------|--------|-------|-------------|
| Average execution time | ~7.0 seconds | ~3.5 seconds | ~50% |
| Total function calls | ~40 million | ~24 million | ~40% |
| _make_hashable calls | ~1.8 million | ~98,000 | ~95% |
| get_hash calls | ~145,000 | ~46,500 | ~68% |

## Testing

The changes were tested on a complex template with numerous interconnected conditions. The template contains approximately 50+ conditions with complex dependencies between them, creating a challenging validation scenario. Multiple benchmark runs confirmed consistent performance improvements with no change in validation results.

## Additional Context

This optimization is particularly impactful for templates with complex conditions and nested resources. The approach maintains the same validation behavior while significantly reducing computational overhead.

## Future Considerations

While this PR makes significant improvements, there are additional optimization opportunities:
1. Further optimizing the satisfiability checks in condition evaluation
2. Improving JSON schema validation performance
3. Considering a more fundamental redesign that separates condition status evolution from instance data

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.